+++ .gitignore
@@ -2,5 +2,6 @@
 /src/.*
 /src/__pycache__
 /src/utils/__pycache__
+/src/utils/models/__pycache__
 /data/*.pth
 test.ipynb
