+++ src/data/images/transform_truck.png
