+++ src/utils/models/ConvNet_1_model.py
@@ -1,0 +1,27 @@
+import torch.nn as nn
+import torch.nn.functional as F
+
+class ConvNet_1(nn.Module):
+    """
+    A simple Convolutional Neural Network with two convolutional layers,
+    followed by three fully connected layers for classification. This model 
+    uses ReLU activations and max pooling after each convolution.
+    """
+
+    def __init__(self):
+        super(ConvNet_1, self).__init__()
+        self.conv1 = nn.Conv2d(3, 6, 5)
+        self.pool = nn.MaxPool2d(2, 2)
+        self.conv2 = nn.Conv2d(6, 16, 5)
+        self.fc1 = nn.Linear(16 * 5 * 5, 120)
+        self.fc2 = nn.Linear(120, 84)
+        self.fc3 = nn.Linear(84, 10)
+
+    def forward(self, x):
+        x = self.pool(F.relu(self.conv1(x)))
+        x = self.pool(F.relu(self.conv2(x)))
+        x = x.view(-1, 16 * 5 * 5)
+        x = F.relu(self.fc1(x))
+        x = F.relu(self.fc2(x))
+        x = self.fc3(x)
+        return x
